+++ src/@locale/components/index.js
@@ -2,6 +2,7 @@ import buttons    from './buttons'
 import dropdowns  from './dropdowns'
 import modal      from './modal'
 import pagination from './pagination'
+import picker     from './picker'
 import splitPanel from './splitPanel'
 
 const components =
@@ -10,6 +11,7 @@ const components =
     dropdowns,
     modal ,
     pagination ,
+    picker ,
     splitPanel ,
 };